requirements.txt
transformers
pillow
pyrogram
torch
opencv-python
uvloop
cryptg
tgcrypto
motor